+++ metaseq/models/transformer_decoder.py
@@ -3,8 +3,8 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-import logging
 import math
+import logging
 from typing import Any, Dict, List, Optional
 
 import torch
@@ -14,15 +14,26 @@ from metaseq.dataclass.constants import UNSPECIFIED_DOC_SEP
 
 from metaseq import utils
 from metaseq.distributed import utils as distributed_utils, fsdp_wrap
-from metaseq.models import IncrementalDecoder
+from metaseq.models import BaseDecoder
 from metaseq.modules import (
     Dropout,
     LayerNorm,
     PositionalEmbedding,
     TransformerDecoderLayer,
+    ModelParallelTransformerDecoderLayer,
     Linear,
 )
 from metaseq.modules.checkpoint_activations import checkpoint_wrapper
+
+try:
+    from megatron import mpu
+    from megatron.mpu import (
+        gather_from_tensor_model_parallel_region,
+    )
+
+    has_megatron_submodule = True
+except (ImportError, ModuleNotFoundError):
+    has_megatron_submodule = False
 
 logger = logging.getLogger(__name__)
 
@@ -53,7 +64,7 @@ def _log_weight_stats(tensor, name):
     )
 
 
-class TransformerDecoder(IncrementalDecoder):
+class TransformerDecoder(BaseDecoder):
     """
     Transformer decoder consisting of *args.decoder_layers* layers. Each layer
     is a :class:`TransformerDecoderLayer`.
@@ -101,6 +112,7 @@ class TransformerDecoder(IncrementalDecoder):
                 learned=args.decoder_learned_pos,
                 learned_sinusoidal=getattr(args, "decoder_learned_sinusoidal", False),
                 full_megatron_init=getattr(args, "full_megatron_init", False),
+                pos_init_scalar=getattr(args, "pos_init_scalar", 1.0),
                 megatron_init_sigma=getattr(args, "megatron_init_sigma", 0.006),
                 truncate_init=getattr(args, "truncate_init", False),
             )
@@ -314,7 +326,7 @@ class TransformerDecoder(IncrementalDecoder):
                 tokens, incremental_state=incremental_state, positions=positions
             )
 
-        # see IncrementalDecoder for important information about
+        # see BaseDecoder for important information about
         # incremental state
         if incremental_state:
             tokens = tokens[:, -1:]
@@ -368,7 +380,7 @@ class TransformerDecoder(IncrementalDecoder):
                 - a dictionary with any model-specific outputs
         """
 
-        # see IncrementalDecoder for important information about
+        # see BaseDecoder for important information about
         # incremental state
         x, extra = self.extract_features(
             prev_output_tokens,
@@ -404,7 +416,7 @@ class TransformerDecoder(IncrementalDecoder):
             prev_output_tokens, token_embeddings, incremental_state
         )
 
-        # see IncrementalDecoder for important information about
+        # see BaseDecoder for important information about
         # incremental state. Note that it may be an empty dictionary.
         if not incremental_state:
             self_attn_mask = self.buffered_future_mask(x, prev_output_tokens)
@@ -496,3 +508,58 @@ class TransformerDecoder(IncrementalDecoder):
             return self._future_mask
         else:
             return self._future_mask[:cur_seq_len, :cur_seq_len]
+
+
+class ModelParallelTransformerDecoder(TransformerDecoder):
+    """
+    Model Parallel Transformer decoder consisting of *args.decoder_layers* layers. Each layer
+    is a :class:`ModelParallelTransformerDecoderLayer`.
+    """
+
+    def build_base_decoder_layer(self, args, **kwargs):
+        return ModelParallelTransformerDecoderLayer(args)
+
+    def output_layer(self, features, **kwargs):
+        """Project features to the vocabulary size."""
+        if not self.share_input_output_embed:
+            raise NotImplementedError(
+                "Model parallel training currently requires --share-decoder-input-output-embed"
+            )
+
+        is_sequence_parallel = getattr(self.args, "sequence_parallel", False)
+        if is_sequence_parallel:
+            input_parallel = features
+        else:
+            input_parallel = mpu.copy_to_tensor_model_parallel_region(features)
+
+        # project back to size of vocabulary
+        x = mpu.LinearWithGradAccumulationAndAsyncCommunication.apply(
+            input_parallel,
+            self.output_projection.weight,
+            None,
+            False,  # gradient_accumulation_fusion
+            False,  # async_grad_allreduce
+            is_sequence_parallel,  # sequence_parallel
+        )
+        # Gather output if model is in inference mode (i.e. evallm or generation) cause both are not yet compatible with
+        # parallel vocab embeddings
+        if getattr(self.args, "criterion") != "vocab_parallel_cross_entropy" or getattr(
+            self, "inference", False
+        ):
+            x = gather_from_tensor_model_parallel_region(x).contiguous()
+
+        return x
+
+    # This hook used as proxy for tracking state if model is in eval or generation mode.
+    def make_generation_fast_(self, **unused):
+        self.inference = True
+
+    def forward_embedding(
+        self,
+        *args,
+    ):
+        x, embed, positions = super().forward_embedding(*args)
+        is_sequence_parallel = getattr(self.args, "sequence_parallel", False)
+        if is_sequence_parallel:
+            x = mpu.scatter_to_sequence_parallel_region(x)
+        return x, embed, positions